% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dfs.R
\name{check_colnames}
\alias{check_colnames}
\title{Compare colnames between two dataframes}
\usage{
check_colnames(data_a, data_b, expect)
}
\arguments{
\item{data_a}{dataframe}

\item{data_b}{dataframe}

\item{expect}{c("equal", "not_equal")}
}
\value{
list with result and evaluation
}
\description{
Compare colnames between two dataframes
}
