% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dfs.R
\name{check_n_rows}
\alias{check_n_rows}
\title{Compare number of rows between two dataframes}
\usage{
check_n_rows(data_a, data_b, expect)
}
\arguments{
\item{data_a}{dataframe}

\item{data_b}{dataframe}

\item{expect}{c("equal", "less", "more")}
}
\value{
list with result and evaluation
}
\description{
Compare number of rows between two dataframes
}
