% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dfs.R
\name{compare_dfs}
\alias{compare_dfs}
\title{Compare properties of two dataframes}
\usage{
compare_dfs(
  data_a,
  data_b,
  expect_n_rows = "equal",
  expect_n_cols = "equal",
  expect_colnames = "equal"
)
}
\arguments{
\item{data_a}{dataframe}

\item{data_b}{dataframe}

\item{expect_n_rows}{c("equal", "less", "more"). Defaults to equal.}

\item{expect_n_cols}{c("equal", "less", "more"). Defaults to equal.}

\item{expect_colnames}{c("equal", "not_equal"). Defaults to equal.}
}
\value{
tibble with result and evaluation of checks
}
\description{
Creates a comparison report based on the dimensional properties of the two
dataframes and evaluates this based on your your expectation. The default
expectation is that all properties are equal.
}
\examples{
compare_dfs(mtcars,
            mtcars |> dplyr::filter(cyl != 6))

compare_dfs(mtcars,
            dplyr::select(mtcars, -hp),
            expect_n_cols = "less",
            expect_colnames = "not_equal")

}
