% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowtracker-package.R
\docType{package}
\name{flowtracker-package}
\alias{flowtracker}
\alias{flowtracker-package}
\title{flowtracker: Compare dataframes on different properties}
\description{
This package makes it possible to compare dataframes on different properties (nr of columns, nr of rows, colnames) and to evaluate if this meets your expectation. This enables insight into the data tranformation process. It creates a possibility to check that the transformation gives the expected result and no unexpected behavior is introduced.
}
\keyword{internal}
